Extract user blocking into src/core/blocking.ts
Move blocked-user storage, denial counting, and moderator block/unblock
actions into a dedicated module. Typecheck, 167 tests, and build all
pass.
